e2e/pyproject.toml
[project]
name = "e2e"
version = "0.1.0"
description = "End-to-end tests for h4kstream"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pyjwt>=2.10.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "redis>=5.2.1",
    "types-redis>=4.6.0.20241004",
]

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
follow_imports = "silent"
warn_redundant_casts = true
disallow_any_generics = false
check_untyped_defs = true

[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint]
extend-select = [
    "F", "W", "I", "UP"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
